Override install command in .travis.yaml
We're hitting the issue described here:
https://docs.travis-ci.com/user/languages/javascript-with-nodejs#travis-ci-supports-yarn

Travis tries to use yarn because we have a yarn.lock checked in to
support yarn users. For some reason yarn doesn't build on Travis,
which should probably be looked at in itself, but this updates tells
Travis to use NPM instead - which will hopefully get the build
working.